+++ src/react-components/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease.  When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ## PRERELEASE
 ### Changed
-- added `noLinkElement` to `EditionCard` to recieve an element
+- added `noLinkElement` to `EditionCard` to receive an element
 
 ## [0.0.11] — 2020-01-23
 ### Added